Move integration tests to a separate workspace crate to avoid MSRV issues with test dependencies
