+++ .github/workflows/benchmark.yml
@@ -27,10 +27,10 @@ jobs:
           cache-target: release
           bins: cargo-codspeed
       - name: Build the benchmark target(s)
-        run: cargo codspeed build -p benchmarks
+        run: cargo codspeed build -p benchmarks --features all_features
       - name: Run the benchmarks
         uses: CodSpeedHQ/action@v1
         with:
           run: cargo codspeed run
-          working-directory: palette # mimics how tests work, so the colors CSV can be found
+          working-directory: integration_tests # mimics how tests work, so the colors CSV can be found
           token: ${{ secrets.CODSPEED_TOKEN }}
