+++ .github/workflows/ci.yml
@@ -36,6 +36,19 @@ jobs:
         shell: bash
         working-directory: palette
         run: bash ../scripts/test_features.sh
+  integration_tests:
+    name: integration tests
+    strategy:
+      matrix:
+        os: [ubuntu-latest, macos-latest, windows-latest]
+    runs-on: ${{ matrix.os }}
+    env:
+      RUSTFLAGS: -D warnings
+    steps:
+      - uses: actions/checkout@v3
+      - uses: dtolnay/rust-toolchain@stable
+      - name: Run integration tests
+        run: cargo test -v -p integration_tests --all-features
   check_stable_beta_nightly:
     name: Quick check
     strategy:
@@ -86,6 +99,7 @@ jobs:
     if: success()
     needs:
       - compile_and_test_msrv
+      - integration_tests
       - check_stable_beta_nightly
       - no_std
       - miri
