+++ .github/workflows/coverage.yml
@@ -24,7 +24,7 @@ jobs:
       - uses: dtolnay/rust-toolchain@nightly
       - uses: taiki-e/install-action@cargo-llvm-cov
       - name: Collect code coverage
-        run: cargo +nightly llvm-cov --all-features -p palette --codecov --doctests --output-path codecov.json
+        run: cargo +nightly llvm-cov --all-features -p integration_tests -p palette --codecov --doctests --output-path codecov.json
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v3
         with:
