+++ benchmarks/benches/rgb.rs
@@ -10,7 +10,7 @@ type LinHsv = Hsv<encoding::Linear<encoding::Srgb>>;
 type LinHsl = Hsl<encoding::Linear<encoding::Srgb>>;
 type LinHwb = Hwb<encoding::Linear<encoding::Srgb>>;
 
-#[path = "../../palette/tests/convert/data_color_mine.rs"]
+#[path = "../../integration_tests/tests/convert/data_color_mine.rs"]
 #[allow(dead_code)]
 mod data_color_mine;
 use data_color_mine::{load_data, ColorMine};
