+++ integration_tests/tests/color_checker.rs
