+++ integration_tests/tests/color_checker_data/babel.csv
