+++ integration_tests/tests/color_checker_data/color_checker.csv
