+++ integration_tests/tests/color_checker_data/load_data.rs
