+++ integration_tests/tests/convert/data_color_mine.csv
