+++ integration_tests/tests/pointer_dataset/pointer_data.rs
@@ -148,7 +148,6 @@ pub fn run_from_lab_tests() {
     }
 }
 
-#[cfg(feature = "wide")]
 pub mod wide_f64x2 {
     use super::*;
