+++ palette/Cargo.toml
@@ -40,10 +40,6 @@ std = ["approx?/std"]
 [lib]
 bench = false
 
-[[example]]
-name = "issue_283"
-path = "regression_tests/issue_283.rs"
-
 [dependencies]
 palette_derive = { version = "0.7.3", path = "../palette_derive" }
 fast-srgb8 = "1.0.0"
@@ -76,14 +72,9 @@ optional = true
 default-features = false
 
 [dev-dependencies]
-csv = "1"
-lazy_static = "1"
-serde = "1"
-serde_derive = "1"
 serde_json = "1"
 ron = "=0.8.0" # Pinned due to MSRV mismatch
 enterpolation = "0.2.0"
-scad = "1.2.2" # For regression testing #283
 
 [dev-dependencies.clap]
 version = "3.2.25"
